Move model files to allow coverage exclusion
