+++ src/main/java/uk/gov/companieshouse/kafka_email/EmailKafkaProducer.java
@@ -9,6 +9,7 @@ import uk.gov.companieshouse.kafka.producer.ProducerConfig;
 import uk.gov.companieshouse.kafka.producer.CHKafkaProducer;
 import uk.gov.companieshouse.kafka.serialization.AvroSerializer;
 import uk.gov.companieshouse.kafka.serialization.SerializerFactory;
+import uk.gov.companieshouse.kafka_email.model.EmailSend;
 
 @Service
 public class EmailKafkaProducer {
