--- src/test/java/uk/gov/companieshouse/kafka_email/EmailSendTest.java
@@ -1,5 +1,0 @@
-package uk.gov.companieshouse.kafka_email;
-
-public class EmailSendTest {
-    
-}
